fix: build error

+++ src/domain/usecases/candidates/delete_cv_usecase.ts
@@ -1,14 +1,12 @@
 import User from '#models/user'
 import UploadFileUsecase from '#domain/usecases/files/upload_file_usecase'
 import { inject } from '@adonisjs/core'
-import UserRepository from '#domain/contracts/repositories/user_repository'
 import CandidateRepository from '#domain/contracts/repositories/candidate_repository'
 
 @inject()
 export default class DeleteCvUsecase {
   constructor(
     private readonly uploadFileUsecase: UploadFileUsecase,
-    private readonly userRepository: UserRepository,
     private readonly candidateRepository: CandidateRepository
   ) {}
 
@@ -16,6 +14,10 @@ export default class DeleteCvUsecase {
     const candidate = await this.candidateRepository.findCandidateByUserId(user.id)
 
     // allow to delete the file
+    if (!candidate.cv) {
+      return
+    }
+
     this.uploadFileUsecase.delete(candidate.cv)
     candidate.cv = ''
